dodano dunkcje w pliku decision.py
